check if there are actually contestants
git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@99532 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ extensions/Contest/includes/Contest.class.php
@@ -409,8 +409,11 @@ class Contest extends ContestDBObject {
 				$contestantIds[] = $contestant->getId();
 			}
 			
-			$success = ContestComment::s()->delete( array( 'contestant_id' => $contestantIds ) ) && $success;
-			$success = ContestVote::s()->delete( array( 'contestant_id' => $contestantIds ) ) && $success;
+			if ( count( $contestantIds ) > 0 ) {
+				$success = ContestComment::s()->delete( array( 'contestant_id' => $contestantIds ) ) && $success;
+				$success = ContestVote::s()->delete( array( 'contestant_id' => $contestantIds ) ) && $success;				
+			}
+			
 			$success = ContestContestant::s()->delete( $condition ) && $success;
 		}
